Adds gracedocMD - generates markdown not HTML
This commit adds gracedocMD.grace. This is a modified version of
gracedoc.grace that generates markdown, instead of the original HTML
page. Note that gracedocMD is setup to generate individual
markdown pages - which are tied together with a YAML sidebar
(intended for use with Jekyll's documentation theme). This
functionality can be easily modified or removed.